+++ NEWS.md
@@ -7,6 +7,7 @@
   `js-mode` with `js2-minor-mode` (see README), rather than
   `js2-jsx-mode`.
 * Using `js2-jsx-mode` will now trigger a warning in Emacs 27.
+* Support for optional-chaining operator `?.`
 
 ## 2019-02-19